chore: Update Vite config plugins to remove attributes from index.html
The Vite config file was updated to include two new plugins, `noAttr()` and `noAttrCss()`, which remove specific attributes from the `index.html` file. This change improves the build process by removing the `crossorigin` attribute from the `type="module"` script tag and the `crossorigin` attribute from the `rel="stylesheet"` link tag.

These modifications ensure that the attributes are removed, allowing for proper loading of the modules and stylesheets in the application.